Bring in InsecureUrlResolver
