+++ components/experimental/diceroller-standalone/src/app/loader.ts
@@ -4,6 +4,7 @@
  */
 
 import { BaseHost, IBaseHostConfig } from "@fluidframework/base-host";
+import { IRequest } from "@fluidframework/component-core-interfaces";
 import {
     IFluidModule,
     IFluidPackage,
@@ -13,16 +14,72 @@ import {
     isFluidPackage,
 } from "@fluidframework/container-definitions";
 import { Container } from "@fluidframework/container-loader";
+import {
+    IFluidResolvedUrl,
+    IResolvedUrl,
+    IUrlResolver,
+} from "@fluidframework/driver-definitions";
+import {
+    ITokenClaims,
+    IUser,
+} from "@fluidframework/protocol-definitions";
 import { RouterliciousDocumentServiceFactory, DefaultErrorTracking } from "@fluidframework/routerlicious-driver";
 import { extractPackageIdentifierDetails } from "@fluidframework/web-code-loader";
+import jwt from "jsonwebtoken";
 // eslint-disable-next-line import/no-internal-modules
 import uuid from "uuid/v4";
-import { InsecureUrlResolver } from "./insecureUrlResolver";
+
+class InsecureUrlResolver implements IUrlResolver {
+    constructor(
+        private readonly ordererUrl: string,
+        private readonly storageUrl: string,
+        private readonly tenantId: string,
+        private readonly tenantKey: string,
+        private readonly user: IUser,
+        private readonly documentId: string,
+    ) { }
+
+    public async resolve(request: IRequest): Promise<IResolvedUrl> {
+        const encodedTenantId = encodeURIComponent(this.tenantId);
+        const encodedDocId = encodeURIComponent(this.documentId);
+
+        const documentUrl = `fluid://${new URL(this.ordererUrl).host}/${encodedTenantId}/${encodedDocId}`;
+        const deltaStorageUrl = `${this.ordererUrl}/deltas/${encodedTenantId}/${encodedDocId}`;
+        const storageUrl = `${this.storageUrl}/repos/${encodedTenantId}`;
+
+        const response: IFluidResolvedUrl = {
+            endpoints: {
+                deltaStorageUrl,
+                ordererUrl: this.ordererUrl,
+                storageUrl,
+            },
+            tokens: { jwt: this.auth(this.tenantId, this.documentId) },
+            type: "fluid",
+            url: documentUrl,
+        };
+        return response;
+    }
+
+    public async getAbsoluteUrl(resolvedUrl: IResolvedUrl, relativeUrl: string): Promise<string> {
+        throw new Error("getAbsoluteUrl not implemented");
+    }
+
+    private auth(tenantId: string, documentId: string) {
+        const claims: ITokenClaims = {
+            documentId,
+            scopes: ["doc:read", "doc:write", "summary:write"],
+            tenantId,
+            user: this.user,
+        };
+
+        return jwt.sign(claims, this.tenantKey);
+    }
+}
 
 class WebpackCodeResolver implements IFluidCodeResolver {
     constructor(private readonly port: number) { }
     async resolveCodeDetails(details: IFluidCodeDetails): Promise<IResolvedFluidCodeDetails> {
-        const baseUrl = details.config.cdn ?? `http://localhost:${this.port}`;
+        const baseUrl = `http://localhost:${this.port}`;
         let pkg = details.package;
         if (typeof pkg === "string") {
             const resp = await fetch(`${baseUrl}/package.json`);
